Move scripts into autoconfig package and add pyproject
