+++ README.md
@@ -4,7 +4,7 @@ This project demonstrates how to collect system facts from hosts using Ansible
 and render the results in a small React application. The playbook gathers the
 list of system users, listening network ports and several basic metrics such as
 CPU load, memory usage and disk space. The collected data is saved as JSON. A
-Python helper script located in the `scripts/` directory runs the playbook and
+Python helper script located in the `autoconfig` package runs the playbook and
 generates a web page using templates stored in `templates/`.
 
 ## Requirements
@@ -32,6 +32,14 @@ npm run build
 ```
 The last two commands generate `templates/dist/bundle.js` used by the HTML template.
 
+### Build and install
+Create a wheel to install the helper tools as a package:
+
+```bash
+python -m build
+pip install dist/*.whl
+```
+
 ## Usage
 Run the helper script which will invoke the playbook, generate a small React
 site in the `results` directory and launch a local nginx server to serve it
@@ -40,7 +48,7 @@ from the local machine only so it can be used for quick testing without a full
 Ansible setup:
 
 ```bash
-python3 scripts/collect_and_visualize.py \
+python3 -m autoconfig.collect_and_visualize \
   --output-dir results \
   --inventory ansible/hosts.ini \
   --port 8080
@@ -50,7 +58,7 @@ To execute the playbook for specific hosts in parallel you can pass them via
 `--hosts`:
 
 ```bash
-python3 scripts/collect_and_visualize.py \
+python3 -m autoconfig.collect_and_visualize \
   --inventory ansible/hosts.ini \
   --hosts host1,host2 \
   --output-dir results
@@ -69,7 +77,7 @@ written to `results/nginx.conf`.
 Alternatively you can run the API directly using:
 
 ```bash
-python3 server.py --port 8000 --results-dir results
+python3 -m autoconfig.server --port 8000 --results-dir results
 ```
 
 The API listens on port 5000 by default and serves host data from the
@@ -81,7 +89,7 @@ Both the helper script and the Flask API use Python's `logging` module. Set
 the `LOG_LEVEL` environment variable to change verbosity, for example:
 
 ```bash
-LOG_LEVEL=DEBUG python3 scripts/collect_and_visualize.py
+LOG_LEVEL=DEBUG python3 -m autoconfig.collect_and_visualize
 ```
 
 ### API Token
@@ -90,7 +98,7 @@ Set the `API_TOKEN` environment variable to protect the `/api/hosts` and
 the same token:
 
 ```bash
-API_TOKEN=secret python3 server.py
+API_TOKEN=secret python3 -m autoconfig.server
 ```
 
 Example request for listing hosts:
@@ -129,7 +137,7 @@ docker build -t autocfg .
 Ensure the `results` directory contains data generated by the helper script:
 
 ```bash
-python3 scripts/collect_and_visualize.py \
+python3 -m autoconfig.collect_and_visualize \
   --output-dir results \
   --inventory ansible/hosts.ini \
   --port 8080
