+++ tests/test_server_reload.py
@@ -4,7 +4,7 @@ import json
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-import server
+from autoconfig import server
 
 
 def _setup(tmp_path, monkeypatch):
